fix: remove code duplication by adding abstract sink class and generic event class
